- fix crash when debugging GLSL geometry shader with lines_adjacency input - some minor fixes after valgrind/memcheck session

+++ src/SHADERed/UI/PipelineUI.cpp
@@ -772,7 +772,7 @@ namespace ed {
 		ImGui::NextColumn();
 		ImGui::PopStyleColor();
 
-		ImGui::Text("%d", els.size());
+		ImGui::Text("%d", (int)els.size());
 		ImGui::NextColumn();
 
 		/* TYPE */
